Apex data type changed; Rising logic changed;

+++ Centaurus.Domain/Quanta/QuantumStorage.cs
@@ -13,7 +13,7 @@ namespace Centaurus.Domain
     {
         static Logger logger = LogManager.GetCurrentClassLogger();
 
-        Dictionary<ulong, MessageEnvelope> storage = new Dictionary<ulong, MessageEnvelope>();
+        Dictionary<long, MessageEnvelope> storage = new Dictionary<long, MessageEnvelope>();
 
         Timer cleanUpTimer;
 
@@ -21,8 +21,8 @@ namespace Centaurus.Domain
         int cleanupInterval = 60;
         int quantumsInMemoryPeriod = 5 * 60 * 1000;
 
-        public ulong CurrentApex { get; private set; } = 0;
-        public QuantumStorage(ulong currentApex)
+        public long CurrentApex { get; private set; } = 0;
+        public QuantumStorage(long currentApex)
         {
             CurrentApex = currentApex;
 #if !DEBUG
@@ -38,9 +38,9 @@ namespace Centaurus.Domain
         {
             try
             {
-                var currentTicks = (ulong)DateTime.UtcNow.Ticks;
+                var currentTicks = DateTime.UtcNow.Ticks;
                 var quantumsToCleanup = storage
-                    .Where(q => new TimeSpan((long)(currentTicks - ((Quantum)q.Value.Message).Timestamp)).TotalMilliseconds > quantumsInMemoryPeriod)
+                    .Where(q => new TimeSpan(currentTicks - ((Quantum)q.Value.Message).Timestamp).TotalMilliseconds > quantumsInMemoryPeriod)
                     .Select(q => q.Key);
 
                 foreach (var q in quantumsToCleanup)
@@ -62,7 +62,7 @@ namespace Centaurus.Domain
             if (Global.IsAlpha)
             {
                 quantum.Apex = ++CurrentApex;
-                quantum.Timestamp = (ulong)DateTime.UtcNow.Ticks;
+                quantum.Timestamp = DateTime.UtcNow.Ticks;
             }
             else if (quantum.Apex == default) //when auditor receives quantum, the quantum should already contain apex
                 throw new Exception("Quantum has no apex");
@@ -71,14 +71,14 @@ namespace Centaurus.Domain
             storage.Add(quantum.Apex, envelope);
         }
 
-        public MessageEnvelope GetQuantum(ulong apex)
+        public MessageEnvelope GetQuantum(long apex)
         {
             if (!storage.ContainsKey(apex))
                 throw new Exception($"Quantum {apex} was not found");
             return storage[apex];
         }
 
-        public IEnumerable<MessageEnvelope> GetAllQuantums(ulong apexCursor = 0)
+        public IEnumerable<MessageEnvelope> GetAllQuantums(long apexCursor = 0)
         {
             if (apexCursor == 0)
                 return storage.Values;
